requirements.txt
# Labeling
torch>=1.9.0
torchvision>=0.10.0
# Core
fastapi==0.115.6
pydantic==2.10.1
uvicorn==0.31.1
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
ultralytics==8.3.28
transformers==4.46.2
opencv-python==4.10.0.84
pillow==11.0.0
requests==2.32.3
# Validation & Testing
pytest==8.3.3
scikit-learn==1.7
torcheval==0.0.7
matplotlib==3.9.2
numpy==2.1.1
# Logging & Monitoring
loguru==0.7.2
prometheus-fastapi-instrumentator==2.0.1